Move React-Flow into Flowchart component
